Simplify Rule addition
