+++ README.md
@@ -14,7 +14,7 @@
 # Objectives of the package
 Deploy a compliance-as-code engine to provide insights on the compliance status of AWS accounts and resources in a multi-account environment. In addition of the engine, several sets of Rules (or RuleSets) can be deployed and are customizable depending on your environment.
 
-## The 10 Key Features
+## The 9 Key Features
 1. Analyze current situation and trends from the compliance account as all data are pushed in a Datalake.
 2. Use your favorite analytics tool (Amazon QuickSight, Tableau, Splunk, etc.) as the data is formatted to be directly consumable.
 3. Classify your AWS accounts by sensitivity.
@@ -22,9 +22,8 @@ Deploy a compliance-as-code engine to provide insights on the compliance status 
 5. Store all historical data of all the changes by storing the compliance record in a centralized and durable Amazon S3 bucket.
 6. Deploy easily in dozens of accounts: by having a 1-step process for any new application account via AWS CloudFormation.
 7. Protect the code base: by centralizing the code base of all the compliance-as-code rules in a “compliance account”.
-8. Enable a cost-effective approach by limiting the number of individual AWS Config Rules, when required.
-9. Make use of the AWS Config Rules Dashboard to display the details of compliance status of your AWS resources by setting up Config Aggregator. 
-10. Notify on non-compliant resources by triggering an Amazon SNS topic.
+8. Make use of the AWS Config Rules Dashboard to display the details of compliance status of your AWS resources by setting up Config Aggregator. 
+9. Notify on non-compliant resources by triggering an Amazon SNS topic.
 
 ## Demo
 See a demo there: https://youtu.be/VR_4209ewIo?t=40m
@@ -72,7 +71,7 @@ Certain resources may have a business need to not follow a particular rule. You 
 
 To add a resource in the whitelist:
 1. Update the file compliance-account-rulesets-setup/compliance-whitelist.json (for model, there are dummy examples).
-2. Ensure that the location of the whitelist is correct in the code of the lambda function named SecurityEpic7-Compliance_Validation
+2. Ensure that the location of the whitelist is correct in the code of the lambda function named COMPLIANCE_RULESET_LATEST_INSTALLED
 
 The resource will be then be noted as COMPLIANT, and the flag "WhitelistedComplianceType" will be set to "True" for traceability.
 
@@ -117,7 +116,7 @@ In the Resource section, add the Config rule (examples provided in the sample Ru
 
 1. Upload in the S3 bucket you initialy created (Initial deployment - 1.), the following:
 - compliance-account-ruleset-setup.yaml
-- new-rule.zip
+- rule-code/new-rule.zip
 
 Note: the name of the code file must be the same as the zipped file, except the extension. The name defined in the yaml templates modified in Step 1 and Step 2 must be as well the same
